requirements.txt
streamlit
requests
python-dotenv
pandas==1.3.5